fix(image): fix the image preview bugs
