+++ lib/constans/presence_constant.dart
@@ -1,0 +1,7 @@
+import 'package:latlong2/latlong.dart';
+
+class PresenceConstants {
+  static const double RADIUS_IN_METERS = 100.0;
+  static const double VISUAL_RADIUS = 50.0;
+  static const LatLng SMKN4_LOCATION = LatLng(-7.98982051550948, 112.6273050048464);
+}
