+++ lib/views/history_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:student_absence/widgets/bottom_nav_bar.dart';
+import 'package:student_absence/widgets/BottomNavbar/bottom_nav_bar.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 class HistoryScreen extends StatelessWidget {
