+++ lib/views/home_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:fl_chart/fl_chart.dart';
-import 'package:student_absence/widgets/bottom_nav_bar.dart';
+import 'package:student_absence/widgets/BottomNavbar/bottom_nav_bar.dart';
 
 class HomeScreen extends StatelessWidget {
   @override
