+++ lib/views/splash_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart'; // Pastikan untuk import GetX
 import 'package:google_fonts/google_fonts.dart';
-import 'package:student_absence/screens/onboarding_screen.dart'; // Import OnboardingScreen
+import 'package:student_absence/views/onboarding_screen.dart'; // Import OnboardingScreen
 
 class SplashScreen extends StatefulWidget {
   @override
